Contracts refactor phase 2: extract ContractTemplateDeployFormComponent
Moves the Create/Deploy tab (template picker, field form, whitelist
management, custom-JSON path, deploy submission, indexer-status
polling) out of ContractsPageComponent into its own component.

Also introduces two shared pieces uncovered during extraction, since
duplicating them would have let two copies drift apart:
- CovenantTemplateService: template-patch-context resolution and
  self-custody-vault arg-building, used by the deploy form, the
  indexer-import preview, and the action-execution engine alike.
- crypto.util.ts: the blake2b/hex helpers, used by both the deploy
  form and the shell's still-in-place registry/alias code.

Removes the dead 'templates' tab (confirmed unreachable in a prior PR
review — no nav button or code path ever set activeTab to it), since
it referenced dozens of the signals being relocated here.

Duplicates (rather than moves) the deploy-tab's SCSS into the new
component, including the per-template --tpl-accent color generator —
Angular's view encapsulation scopes CSS per-component, so the shell's
stylesheet no longer reaches markup rendered by a child. Shell's
scoped copies are left in place, now inert there, to avoid any risk
to the tabs still living in the shell.

Shell: 8,400 -> 6,938 lines (.ts), 3,803 -> 2,780 lines (.html).
