+++ src/app/v2/pages/app/flows/contracts/contracts-page.models.ts
@@ -122,3 +122,5 @@ export type ActionIndexerState = {
   status: 'checking' | 'indexed' | 'not-indexed' | 'unavailable';
   message: string;
 };
+
+export const SELF_CUSTODY_WHITELIST_CAPACITY = 10;
